Move SpaceButtons.tsx into Header
